fix(e2e): point 5 helpers' seed-uuid imports at the consolidated module
Commit 0d7ce0d migrated 5 spec-level imports from `./helpers/seed-ids`
to `./fixtures/seed-uuids` and deleted the old helper file, but
overlooked that 5 helper modules in `apps/web/e2e/helpers/` ALSO import
the constants — via the shorter relative path `./seed-ids` (same dir).
My grep used `helpers/seed-ids` as the search key and missed those.

Local `tsc --noEmit` did not flag this either because
`apps/web/tsconfig.app.json` only includes `src` — `e2e/` is parsed by
Playwright at test time, not by tsc. CI run 25261843978 surfaced the
broken imports as `Cannot find module './seed-ids'` from
`students.ts`, `teachers.ts`, `constraints.ts`, `subjects.ts`,
`users.ts` (and transitively every spec that imports them).

Fix: rewrite to `from '../fixtures/seed-uuids'` in all 5 helpers.

Verified locally with `pnpm exec playwright test --list` — all 386 tests
across 69 files resolve with exit 0.
