+++ apps/web/e2e/helpers/teachers.ts
@@ -7,7 +7,7 @@
  */
 import { expect, type APIRequestContext } from '@playwright/test';
 import { getAdminToken } from './login';
-import { SEED_SCHOOL_UUID } from './seed-ids';
+import { SEED_SCHOOL_UUID } from '../fixtures/seed-uuids';
 
 export const TEACHER_API =
   process.env.E2E_API_URL ?? 'http://localhost:3000/api/v1';
