+++ apps/web/e2e/helpers/users.ts
@@ -16,7 +16,7 @@
  */
 import { expect, type APIRequestContext } from '@playwright/test';
 import { getAdminToken } from './login';
-import { SEED_SCHOOL_UUID } from './seed-ids';
+import { SEED_SCHOOL_UUID } from '../fixtures/seed-uuids';
 
 export const USER_API = process.env.E2E_API_URL ?? 'http://localhost:3000/api/v1';
 export const USER_PREFIX = 'E2E-USR-';
